+++ pkgs/intl4x/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `calendar.dart` entrypoint exposing `Calendar` and `Weekday` with `Weekday.firstDayOfWeek` API.
 - Add more `DateTimeFormat`s.
+- Update `PluralRules.select` to select from plural form options.
 
 ## 1.0.0-alpha.2